feat: 리스트 순서이동 api 완성 , indexOrder로 정렬

+++ src/modules/list/list.service.ts
@@ -19,11 +19,27 @@ export class ListService {
 
   async createList(createListDto: CreateListDto, boardId: number) {
     const list = { ...createListDto, boardId };
-    const newList = await this.listRepository.save(list);
+
+    const lastList = await this.listRepository.findOne({
+      where: { boardId },
+      order: { orderIndex: "DESC" },
+    });
+    // lists를 찾아서 orderIndex가 없으면 생성 // 있으면 orderIndex의 가장 큰 숫자 + 1
+    const newOrderIndex = lastList ? lastList.orderIndex + 1 : 0;
+    const newList = await this.listRepository.save({
+      ...list,
+      orderIndex: newOrderIndex,
+    });
 
     return newList;
+    // 1. order index 를 내림차순으로 가장 앞에 있는 놈을 앞으로 끄집어낸다 (findOnd 에서 order 내림차순 후 return)
+    // 2. 1번에서 나온 놈이 모든 리스트 중에서 제일 큰놈임
+    // 3. 2번에서 나온 놈의 order index 를 + 1 해준다
+    // 4. 3번에서 나온 놈을 저장한다
+    // 5. 아무 리스트도 없을 때 예외처리 해줘야 되고
   }
-
+  //orderIndex 기준으로 내림차순으로 정렬하면 가장 첫번째에 있는 (find(1) list의 orderIndex를 조회하여 새로 생성되는 list의 orderIndex에 +1 해주고
+  //리스트가 하나도 없다면 default값으로 0을 준다
   async findOneList(id: number): Promise<ListsEntity> {
     const list = await this.listRepository.findOne({
       where: { id },
@@ -34,6 +50,7 @@ export class ListService {
     }
     list.card = await this.cardRepository.find({
       where: { listId: list.id },
+      order: { orderIndex: "ASC" },
       select: ["content"],
     });
     return list;
@@ -42,7 +59,7 @@ export class ListService {
   async findAllLists(): Promise<ListsEntity[]> {
     return await this.listRepository.find({
       order: {
-        createdAt: "ASC",
+        orderIndex: "ASC",
       },
     });
   }
@@ -58,21 +75,32 @@ export class ListService {
     return updateList;
   }
 
+  //삭제시 orderIndex순으로 다시 초기화.
   async deleteList(id: number) {
-    return await this.listRepository.delete({ id });
+    const findList = await this.listRepository.findOne({ where: { id } });
+    if (!findList) {
+      throw new NotFoundException(MESSAGES.LIST.NOT_EXISTS);
+    }
+    await this.listRepository.delete({ id });
+    const newlists = await this.listRepository.find({ order: { orderIndex: "ASC" } });
+    for (let i = 0; i < newlists.length; i++) {
+      newlists[i].orderIndex = i;
+      await this.listRepository.save(newlists[i]);
+    }
   }
 
+  // 리스트 순서이동
   async updateOrderList(listIdIndex: number, updateListOrderDto: UpdateListOrderDto) {
     const { newPositionId } = updateListOrderDto;
 
     const lists = await this.listRepository.find({
-      order: { id: "ASC" },
+      order: { orderIndex: "ASC" },
     });
 
-    for (let i = 0; i < lists.length; i++) {
-      lists[i].orderIndex = i;
-      await this.listRepository.save(lists[i]);
-    }
+    // for (let i = 0; i < lists.length; i++) {
+    //   lists[i].orderIndex = i;
+    //   await this.listRepository.save(lists[i]);
+    // }
 
     const currentIndex = lists[listIdIndex];
 
@@ -86,11 +114,11 @@ export class ListService {
     }
 
     // orderIndex를 기준으로 정렬된 리스트 반환
-    const updatedLists = await this.listRepository.find({
-      order: { orderIndex: "ASC" },
-    });
+    // const updatedLists = await this.listRepository.find({
+    //   order: { orderIndex: "ASC" },
+    // });
 
-    return updatedLists;
+    return lists;
   }
 
   // async updateOrderList(listIdIndex: number, updateListOrderDto: UpdateListOrderDto) {
